Refactor news data client and improve request handling
- Removed the deprecated newsrequest.go file and consolidated article and source request handling into a more streamlined structure.
- Introduced a new articleRequest type for managing article queries, enhancing clarity and usability.
- Updated NewsdataClient methods to utilize the new request structure, improving method signatures and overall API consistency.
- Enhanced error handling and validation across article and source requests, ensuring robust input management.
- Revised README.md to reflect the new request structure and provide updated usage examples for fetching articles and sources.
- Improved logging capabilities for better debugging and tracking of requests.
